pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codebase-rag"
version = "0.1.0"
description = "RAG system for codebases using ChromaDB, Langchain, and tree-sitter"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Developer"}
]
keywords = ["rag", "codebase", "chroma", "langchain", "tree-sitter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "chromadb>=0.4.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.2",
    "tree-sitter>=0.20.0,<0.22",
    "tree-sitter-languages",
    "click>=8.0.0",
    "tqdm>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
rag = "codebase_rag.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["codebase_rag*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
